[SKP-102] feat : category Selector 컴포넌트 수정
